refactor: convert `apps/meteor/app/apps/server/converters/uploads.js` to TypeScript
